backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
bcrypt==4.1.2
SQLAlchemy==2.0.28
mysqlclient==2.2.4
PyMySQL==1.1.0
